Minor fixes

+++ include/Constants.h
@@ -9,20 +9,20 @@
 //#include "CutsWZ.h"
 
 
-const float lumi_fb_blind = 1.371;
-const float lumi_fb       = 2.318;
-const float lumi_error_percent = 2.7;
+const double lumi_fb_blind = 1.371;
+const double lumi_fb       = 2.318;
+const double lumi_error_percent = 2.7;
 
 const int ELECTRON_FLAVOUR = 11;
 const int MUON_FLAVOUR     = 13;
 const int TAU_FLAVOUR      = 15;
 const int Z_FLAVOUR        = 23;
 
-const float ELECTRON_MASS =   0.000511;  // [GeV]
-const float MUON_MASS     =   0.106;     // [GeV]
-const float TAU_MASS      =   1.777;     // [GeV]
-const float Z_MASS        =  91.188;     // [GeV]
-const float H_MASS        = 125.0;       // [GeV]
+const double ELECTRON_MASS =   0.000511;  // [GeV]
+const double MUON_MASS     =   0.106;     // [GeV]
+const double TAU_MASS      =   1.777;     // [GeV]
+const double Z_MASS        =  91.188;     // [GeV]
+const double H_MASS        = 125.0;       // [GeV]
 
 
 // Systematic uncertainties
@@ -87,16 +87,16 @@ const Color_t color_HZ        = kOrange+1;
 
 // Branching ratios
 //------------------------------------------------------------------------------
-const float W2e     = 0.1075;
-const float W2m     = 0.1057;
-const float W2tau   = 0.1125;
-const float Z2ll    = 0.033658;
-const float WZ23lnu = 3 * Z2ll * (W2e + W2m + W2tau);
+const double W2e     = 0.1075;
+const double W2m     = 0.1057;
+const double W2tau   = 0.1125;
+const double Z2ll    = 0.033658;
+const double WZ23lnu = 3 * Z2ll * (W2e + W2m + W2tau);
 
 
 const int nmetbin = 5;
 
-const float metbins[nmetbin+1] = {0, 20, 25, 30, 45, 3000};  // [GeV]
+const double metbins[nmetbin+1] = {0, 20, 25, 30, 45, 3000};  // [GeV]
 
 
 const int njetbin = 3;  // 0jet, 1jet, 2+jet, all
@@ -142,16 +142,16 @@ const TString lchannel[nchannel] = {
 
 // https://twiki.cern.ch/twiki/bin/view/CMS/BtagRecommendation76X
 // pfCombinedMVAV2BJetTags
-const float cMVAv2L = -0.715;
-const float cMVAv2M =  0.185;
-const float cMVAv2T =  0.875;
+const double cMVAv2L = -0.715;
+const double cMVAv2M =  0.185;
+const double cMVAv2T =  0.875;
 
 
 // https://twiki.cern.ch/twiki/bin/view/CMS/BtagRecommendation76X
 // pfCombinedInclusiveSecondaryVertexV2BJetTags
-const float CSVv2L = 0.460;
-const float CSVv2M = 0.800;
-const float CSVv2T = 0.935;
+const double CSVv2L = 0.460;
+const double CSVv2M = 0.800;
+const double CSVv2T = 0.935;
 
 
 // MET filters
